Back-port of #8691 into 74X. Could be extremely useful also for the 0T scenario.