@cmsbuild 2015-05-27
The tests are being triggered in jenkins.