@slava77 2015-05-27
Hi Marco, just making a blind comparison of the number of lines, this PR doesn't match #8691 
what are the changes?