@rovere 2015-05-27
@slava77 you are right, I might have missed 1 commit from #8691 
I'll update this PR accordingly.
Thanks for noticing.