@slava77 2015-05-27
ah, ok, good that I checked.
At brief inspection, it looks like the changes in tracking parameters in the *iteration*cff are missing