@slava77 2015-05-27
@cmsbuild please test