@cmsbuild 2015-05-27
+1
Tested at: 20e3c43dae83a15c4db7de6d2e4bd2ef570d1430
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-9296/5064/summary.html