@slava77 2015-05-27
this means that T0 processing scenarios need to be defined.
Just having a customize function is not enough (new PR, I suppose)