@VinInn 2015-05-27
Sure, and I bet a new release will be needed.
Do you not expect other detectors to provide new code as well, do you?