@slava77 2015-05-27
@VinInn 
yes, we are collecting inputs.
you are the first to have a solution for your area.