@cmsbuild 2015-05-27
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-9296/5064/summary.html