@franzoni 2015-05-28
@VinInn @rovere  can you please circulate to hn instructions to activate these changes ?
(I notice that all what you've modified are runI python, which I doubut are part of the default RECO)
Thanks.