@rovere 2015-05-28
@franzoni Ciao Giovanni,
detailed instructions have been provided and attached to #8691 
Before circulating them "publicly" over HN maybe  it's better to see the performances over some RelVal sample@0T (yes, privately re-reco'd with the customise, unless you can offer a central re-reco w/ this customise on)