@slava77 2015-05-28
summary for https://github.com/cms-sw/cmssw/pull/9296	86fe202
- tested locally in CMSSW_7_4_3 /test area sign743b/
- the main test is based on run 200781 lumi=80 (~PU=30 with 0 T from 2012C)  /store/data/Run2012C/MinimumBias/RAW/v1/000/200/781/7ED5DB39-DDE4-E111-9386-003048F118D2.root , using 1000.0 configuration otherwise. 100 events in the test.
- technical performance seems comparable: processing time up from 10.6 to 13.7 s/evt (peak moves from 29.4 to 35.8 s/evt), both excluding the first event dominated by conditions fetch.

Some object comparisons notes
- most tracks don't have MVA value assigned
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_floatedmvaluemap_generaltracks_mvavals_reco_obj_values_](https://cloud.githubusercontent.com/assets/4676718/7857846/d14a5e98-0501-11e5-85de-07d5203c88c6.png)
- there are 50% more tracks (from the following distributions they appear to be "better" tracks)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_recotracks_generaltracks__reco_obj_pt](https://cloud.githubusercontent.com/assets/4676718/7858423/e8c09b10-0505-11e5-9bc3-23acd50e3df4.png)

![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_minrecotracks_generaltracks__reco_obj_normalizedchi2 29](https://cloud.githubusercontent.com/assets/4676718/7858127/cd04063e-0503-11e5-9802-deb88d1c0179.png)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_recotracks_generaltracks__reco_obj_algo](https://cloud.githubusercontent.com/assets/4676718/7858347/95206102-0505-11e5-8f38-36beaad1f47f.png)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_recotracks_generaltracks__reco_obj_eta](https://cloud.githubusercontent.com/assets/4676718/7858366/a45c1454-0505-11e5-86e2-65812ab891c3.png)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_recotracks_generaltracks__reco_obj_found](https://cloud.githubusercontent.com/assets/4676718/7858371/ab7fb0ba-0505-11e5-91e1-16747f7d02e7.png)

- more electron gsf tracks as well
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recogsftracks_electrongsftracks__reco_obj_pt](https://cloud.githubusercontent.com/assets/4676718/7857965/a27626aa-0502-11e5-911a-5dbbf9097169.png)
- but fewer PF egamma objects (more tracks probably kill some of them in the preselection)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recopfcandidates_particleflowegamma__reco_obj_pt](https://cloud.githubusercontent.com/assets/4676718/7857992/c949bc6a-0502-11e5-84fa-815f528fff47.png)

- calo jets change a bit (likely due to the track->hbhe cleaning step)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recocalojets_ak4calojets__reco_obj_et](https://cloud.githubusercontent.com/assets/4676718/7857950/8ceba5c6-0502-11e5-9421-04d6679d66a8.png)

PF reco is not quite trivial to understand (do we even know that it works OK?)
- about 30% less charged hadrons (opposite to 50% more tracks .. looks like PF RECO doesn't like these tracks.)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recopfcandidates_particleflow__reco_obj_pt10](https://cloud.githubusercontent.com/assets/4676718/7858037/284d24e0-0503-11e5-9f85-c52b3255d7d5.png)
- similar decrease in electrons (this should follow what happens in the egamma objects; since the sample is mostly fakes, the meaning of decrease is unclear)
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recopfcandidates_particleflow__reco_obj_pt20](https://cloud.githubusercontent.com/assets/4676718/7858057/513442ee-0503-11e5-9113-93961df1096c.png)

- muons are ~unchanged
- photons are ~unchanged
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recopfcandidates_particleflow__reco_obj_pt40](https://cloud.githubusercontent.com/assets/4676718/7858085/7f9aea0c-0503-11e5-9fd2-e1f26d8134a0.png)
- ~15% more neutral hadrons
![all_sign743b-200781vsorig-200781_runminbias2011awf1000p0c_log10recopfcandidates_particleflow__reco_obj_pt50](https://cloud.githubusercontent.com/assets/4676718/7858091/8c8f7994-0503-11e5-86aa-3420a356548e.png)


Overall, there are indications that tracking is indeed better, but higher level reco downstream (PF, specifically) maybe just not well adjusted to the 0T input track properties.
MC with truth matching would help to understand what's happening.