@slava77 2015-05-28
@bachtis Michalis, do you have any high expectations for PF reco with 0 T?
@lgray what about egamma?