@bachtis 2015-05-28
Hi Slava ,
I have discussed with PC . There are many ideas that are configurable and could be done at PF @0T
Examples: 
[Run PF without tracks ~ 20% improvement for jet resolution wrt CaloJets] 
Link tracks and Calo deposits and create PF candidates with Calo energy and track direction [and try to associate to PV]

Now the request from PC was to not invest in that direction for the high level reconstruction for now since we are looking at a middle term solution.
just for my education: @rovere  @VinInn  what is the pt of the tracks at 0 T    and the error that you give?