@VinInn 2015-05-28
Seeds starts with 5GeV and some sort of large error.
than KalmanFilter is free to adjust most probably taking into account multiple-scattering...